support abs restore
