+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM php:7.1
+FROM php:7.4
 
 ENV DEBIAN_FRONTEND noninteractive
 ENV COMPOSER_ALLOW_SUPERUSER 1
