+++ phpstan.neon
@@ -2,5 +2,4 @@ parameters:
     checkMissingIterableValueType: false
     ignoreErrors:
         - '#Parameter \#4 \$sourceBucketId of method Keboola\\StorageApi\\Client\:\:linkBucket\(\) expects int, string given.#'
-        - '#Parameter \#1 \$message of class Keboola\\StorageApi\\ClientException constructor expects null, string given.#'
-        - '#Parameter \#1 \$bucket of method Keboola\\ProjectRestore\\S3Restore\:\:restoreBucket\(\) expects Keboola\\ProjectRestore\\StorageApi\\BucketInfo#'
+        - '#Parameter \#1 \$message of class Keboola\\StorageApi\\Exception constructor expects null, string given.#'
